Add special handling of bootstrap-file for POST requests
If the Test object has "post" = true, and there is an attribute "bootstrap-file", send a multipart/form-data POST request instead.

The "bootstrap-file" attribute specifies the path to the file that should be sent.

Note: does not handle missing/unable to be accessed files properly yet (HttpClient's logger will spew out a lot of messages)